Fix plan-review findings across ITER_01–ITER_06
Deviations fixed:
- Node.border_width and Edge.width: int → float (default 1.5 is a float)

Errors fixed:
- Remove unused `import math` in Flow.group()
- JS: extract PAN_ON_DRAG to module-level constant (was inline array prop on
  ReactFlow, causing unnecessary stability issues)
- JS: rewrite layout handler — replaced nested setEdges-inside-setNodes
  anti-pattern with direct model.get() reads; now calls pushHistory() so
  layout changes are undoable via the commit seam (ITER_03 requirement)
- serialize.py: fix to_mermaid() arrowclosed arrow construction — was
  producing "---->" (invalid mermaid); removed the buggy rstrip override

Deviation resolved (ITER_06 emit contract):
- Updated figureflow.d.ts: emit is documented as data.emit (xyflow's custom
  node API only exposes user data through the `data` prop; separate top-level
  props aren't supported). Updated NodeProps/EdgeProps interfaces accordingly.
- Updated examples/custom_component.py to destructure emit from data.

Gaps filled (ITER_06 packaging — tests only):
- tests/test_figureflow.py: 44 pytest cases covering Node/Edge.to_dict(),
  positions(), group(), from_json(to_json(x)) round-trip, and to_mermaid()
- pyproject.toml: [dev] optional dep (pytest>=7), pytest testpaths config

+++ src/figureflow/__init__.py
@@ -76,7 +76,7 @@ class Node:
     shape: Shape = field(default_factory=lambda: _BASELINE["shape"])
     fill: str = field(default_factory=lambda: _BASELINE["fill"])
     border_color: str = field(default_factory=lambda: _BASELINE["border_color"])
-    border_width: int = field(default_factory=lambda: _BASELINE["border_width"])
+    border_width: float = field(default_factory=lambda: _BASELINE["border_width"])
     font_family: str = field(default_factory=lambda: _BASELINE["font_family"])
     font_size: int = field(default_factory=lambda: _BASELINE["font_size"])
     font_color: str = field(default_factory=lambda: _BASELINE["font_color"])
@@ -157,7 +157,7 @@ class Edge:
     marker_start: str = field(default_factory=lambda: _BASELINE["marker_start"])
     marker_end: str = field(default_factory=lambda: _BASELINE["marker_end"])
     color: str = field(default_factory=lambda: _BASELINE["edge_color"])
-    width: int = field(default_factory=lambda: _BASELINE["edge_width"])
+    width: float = field(default_factory=lambda: _BASELINE["edge_width"])
     dash: str = field(default_factory=lambda: _BASELINE["dash"])
     font_family: str = field(default_factory=lambda: _BASELINE["font_family"])
     font_size: int = field(default_factory=lambda: _BASELINE["font_size"])
@@ -288,8 +288,6 @@ class Flow(anywidget.AnyWidget):
 
     def group(self, node_ids: List[str], label: str = "") -> str:
         """Create a parent group containing the given nodes; return the group id."""
-        import math
-
         group_n = sum(1 for n in self.nodes if n.get("type") == "group")
         group_id = f"group-{group_n}"
